backend/requirements.txt
fastapi
uvicorn

pytesseract
Pillow
pdfplumber

python-multipart
pydantic

openai
